+++ client/src/scene/BaseScene.cpp
@@ -1,6 +1,7 @@
 #include "BaseScene.h"
 #include "Scene.h"
 #include "Game.h"
+#include "utils.h"
 
 BaseScene::BaseScene(int id, String name)
 :m_Name(name),
@@ -118,7 +119,7 @@ void BaseScene::AddNpc(const char* player_name, float x, float y, int dir, int r
 
 void BaseScene::AddPet(const char* player_name, int x, int y, int dir, int role_id, int action_id)
 {
-	Pet* pet = new Pet(player_name, x, y, dir, role_id, action_id);
+	Pet* pet = new Pet(player_name, (float)x, (float)y, dir, role_id, action_id);
 	m_Players.push_back(pet);
 }
